do not close xsettingsd in e16

+++ usr/lib/user-manager/hooks.d/e17-wizard/last/90-unset-gtk-theme.sh
@@ -1,3 +1,5 @@
 #!/bin/sh
 
-killall xsettingsd 2>/dev/null 1>&2 || true
+if [[ -n "$ESTART" ]] ; then
+    killall xsettingsd 2>/dev/null 1>&2 || true
+fi
